% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{asymp.variance}
\alias{asymp.variance}
\title{asymp.variance}
\usage{
asymp.variance()
}
\arguments{
\item{a}{structure?}
}
\value{
the variance matrix
}
\description{
This function estimates the variance matrix of DINA by the asymptotic variance matrix with reasonably good nuisance function estimators and regularity conditions
}
