% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{base.cumulative.hazard}
\alias{base.cumulative.hazard}
\title{base.cumulative.hazard}
\usage{
base.cumulative.hazard(eta, time, status, time.test = NULL)
}
\arguments{
\item{eta}{vector of linear predictors in a fitted Cox model.}

\item{time}{vector of survival times.}

\item{status}{vector of censored status. \code{Status = 1} stands for not censored, and \code{status = 0} stands for being censored.}

\item{time.test}{vector of time to evaluate the baseline cumulative hazard at.}
}
\value{
vector of baseline cumulative hazard.
}
\description{
This function computes the baseline cumulative hazard for a fitted Cox model.
}
