% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.DINA.R
\name{confint.DINA}
\alias{confint.DINA}
\title{confint.DINA}
\usage{
\method{confint}{DINA}(object, level = 0.95, ...)
}
\arguments{
\item{object}{object of class "DINA".}

\item{level}{the confidence level required. Default is 0.95.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a matrix with columns giving lower and upper confidence limits for each parameter. These will be labeled as (1-\code{level})/2 and 1 - (1-\code{level})/2 in % (by default 2.5% and 97.5%).
}
\description{
This function computes confidence intervals for one or more parameters in a DINA object. The method assumes normality.
}
