% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{cox.inference}
\alias{cox.inference}
\title{cox.inference}
\usage{
cox.inference(eta, time, status, X)
}
\arguments{
\item{eta}{vector of linear predictors in a fitted Cox model.}

\item{time}{vector of survival times.}

\item{status}{vector of censored status. \code{Status = 1} stands for not censored, and \code{status = 0} stands for being censored.}

\item{X}{covariate matrix. Each row stands for an observation, each column stands for a covariate.}
}
\value{
a list of data.
 \itemize{
 \item{U}{the first order derivative of the partial likelihood.}
 \item{I}{the second order derivative of the partial likelihood.}
}
}
\description{
This function computes the first and second order derivatives of the partial likelihood of a Cox model with regard to the regression coefficients.
}
