% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{dataGnr}
\alias{dataGnr}
\title{dataGnr}
\usage{
dataGnr(
  n,
  d,
  alpha,
  beta,
  e = 0,
  theta = NULL,
  delta = 0,
  typeX = "uniform",
  family,
  typeMisMu0 = "quadratic",
  params = list()
)
}
\arguments{
\item{n}{sample size.}

\item{d}{number of covariates.}

\item{alpha}{coefficient vector for the control natural parameter function. Without further specification, the control natural parameter function is the sum of two parts: a linear function of the covariates \code{X} (we include the intercept as the last column of \code{X}) with the coefficient vector \code{alpha}, and a non-linear part determined by the type of misspecification \code{typeMisMu0} and the magnitude \code{delta}.}

\item{beta}{coefficient vector for the heterogeneous treatment effect. Without further specification, the heterogeneous treatment effect is a linear function of the covariates \code{X} (we include the intercept as the last column of \code{X}) with the coefficient vector \code{beta}.}

\item{e}{constant propensity score value. If \code{e} is not zero, then the propensity score is set at \code{e} for all units. Default is zero.}

\item{theta}{coefficient vector for the propensity score. Without further specification, the propensity score follows a logistic regression model of the covariates \code{X} (we include the intercept as the last column of \code{X}) with the coefficient vector \code{theta}.}

\item{delta}{the magnitude of the non-linear part of the control natural parameter function.}

\item{typeX}{distribution of the covariates. Default is the uniform distribution U(0,1).}

\item{family}{family of responses. The response can be one of the following families: "gaussian", "binomial", "poisson", "exponential", "gamma", "cox".}

\item{typeMisMu0}{type of the non-linear part of the control natural parameter function. The non-linear part can be one of the following types: "quadratic", "interaction", "three-interaction", "sin", "cos", "abs", "unobserved", "unobserved linear", "unobserved interaction".}

\item{params}{list of parameters. For censoring, \code{params$censorType} specifies the distribution of the censoring times. The censoring times can be one of the following types: NULL (no censoring), "exponential", "cox", "uniform", "fixed".}
}
\value{
a list of data.
 \itemize{
 \item{X}{covariate matrix.}
 \item{Y}{vector of response.}
 \item{W}{vector of treatment assignment.}
 \item{prop}{vector of propensity score.}
 \item{mu0}{vector of control natural parameter function.}
 \item{tau}{heterogeneous treatment effect.}
}
}
\description{
This function generates the simulated data.
}
\examples{
data.gaussian = dataGnr(n = 100, d = 1, alpha = rep(1,2), beta = rep(0.5,2), e = 0.5, delta = 1, family = "gaussian", typeMisMu0 = "quadratic")
data.binomial = dataGnr(n = 100, d = 5, alpha = rep(1,6), beta = rep(0.5,6), e = 0.5, delta = 1, family = "binomial", typeMisMu0 = "quadratic")
params = list(); params$censorType = "uniform"; params$T0 = 5
data.cox = dataGnr(n = 100, d = 5, alpha = rep(1,6), beta = rep(0.5,6), e = 0.5, delta = 1, family = "cox", typeMisMu0 = "quadratic", params = params)

}
