% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dina-internal.R
\docType{package}
\name{dina-package}
\alias{dina}
\alias{dina-package}
\title{dina: Bayesian Estimation of DINA Model}
\description{
Estimate the Deterministic Input, Noisy "And" Gate (DINA)
cognitive diagnostic model parameters using the Gibbs sampler described
by Culpepper (2015) <doi:10.3102/1076998615595403>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tmsalab/dina}
  \item Report bugs at \url{https://github.com/tmsalab/dina/issues}
}

}
\author{
\strong{Maintainer}: Steven Andrew Culpepper \email{sculpepp@illinois.edu} [copyright holder]

}
