% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DINA.R
\name{DINA.bagging.inference}
\alias{DINA.bagging.inference}
\title{DINA.bagging.inference}
\usage{
DINA.bagging.inference(
  bagging.estimator,
  bootstrap.membership,
  method = "infinitesimal jackknife",
  bias.correction = T
)
}
\arguments{
\item{bagging.estimator}{matrix of DINA estimators. Each row stands for a bootstrap resample, each column stands for a feature.}

\item{bootstrap.membership}{matrix of how many times a unit appears in a bootstrap resample. Each row stands for a bootstrap resample, each column stands for a unit.}

\item{method}{method for estimating the variance matrix. \code{Method} can be a character string naming the estimator: "infinitesimal jackknife", "jackknife". Default is "infinitesimal jackknife".}

\item{bias.correction}{logical value indicating whether to execute bias correction. If \code{TRUE}, an estimator of the bias will be subtracted from the original estimator. Default is \code{TRUE}.}
}
\value{
the variance matrix of bagged DINA estimators.
}
\description{
This function computes the variance matrix of bagged DINA estimators.
}
