% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DINA.R
\name{DINA.bagging}
\alias{DINA.bagging}
\title{DINA.bagging}
\usage{
DINA.bagging(
  data,
  family = "gaussian",
  prop = NULL,
  paramsW = NULL,
  baseline = "gradient boosting",
  params = list(),
  k = 2,
  B = 100,
  variance.method = "infinitesimal jackknife",
  variance.bias.correction = T
)
}
\arguments{
\item{data}{list of data. \code{data$X} denotes the covariate matrix: each row stands for an observation, each column stands for a covariate. \code{data$y} denotes the vector of responses: for \code{family = "cox"}, \code{data$y} is a data.frame with two columns: "time" for the vector of survival times, "status" for the vector of censored status. \code{Status = 1} stands for not censored, and \code{status = 0} stands for being censored. \code{data$W} denotes the vector of treatment assignment: \code{W = 1} stands for treatment, and \code{W = 0} stands for control.}

\item{family}{a character string describing the error distribution and link function to be used in the model. \code{Family} can be one of the following options: "gaussian", "binomial", "poisson", "cox". Default is "gaussian".
#' @param prop method for estimating the propensity score. \code{Prop} can be a character string naming the estimator: "logistic regression", "gradient boosting". \code{Prop} can also be a function and \code{prop(X)} should output the estimated propensity scores for each row of the covariate matrix \code{X}. Default is "logistic regression".}

\item{paramsW}{list of parameters for propensity score estimation.}

\item{baseline}{method for estimating the natural parameter functions in the treatment and control groups. \code{Baseline} can be a character string naming the estimator: "regression", "gradient boosting". \code{Baseline} can also be a list with two members: \code{eta1} denotes the estimator of the natural parameter function in the treatment group,   \code{eta0} denotes the estimator of the natural parameter function in the control group. The functions should take in a covariate matrix \code{X} and output the estimated quantities for each row of the covariate matrix \code{X}. Default is "regression".}

\item{params}{list of parameters for the natural parameter function estimation.}

\item{k}{number of folds for cross-fitting. Default is 2.}

\item{B}{the number of bootstrap resamples. This should be a single positive integer no smaller than 100. Default is 100.}

\item{variance.method}{method for estimating the variance matrix. \code{Method} can be a character string naming the estimator: "infinitesimal jackknife", "jackknife". Default is "infinitesimal jackknife".}

\item{variance.bias.correction}{logical value indicating whether to execute bias correction. If \code{TRUE}, an estimator of the bias will be subtracted from the original estimator. Default is \code{TRUE}.}
}
\value{
an object of class \code{DINA}.
 \itemize{
 \item{estimator}{vector of estimated coefficients.}
 \item{Sigma}{the variance matrix of the coefficients.}
}
}
\description{
This function implements DINA averaged over bootstrap resamples. The estimator DINA is executed over multiple bootstrap resamples, and the final estimator will be the average of the multiple realizations.
}
