% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DINA.R
\name{DINA.fit}
\alias{DINA.fit}
\title{DINA.fit}
\usage{
DINA.fit(
  X,
  Y,
  W,
  family,
  prop,
  paramsW = NULL,
  baseline = "gradient boosting",
  params = list(),
  k = 2,
  index.resample = NULL,
  foldIndex = NULL,
  link.function = NA,
  var.function = NA
)
}
\arguments{
\item{X}{covariate matrix. Each row stands for an observation, each column stands for a covariate.}

\item{Y}{vector of response. For \code{family = "cox"}, \code{Y} is a data.frame with two columns: "time" for the vector of survival times, "status" for the vector of censored status. \code{Status = 1} stands for not censored, and \code{status = 0} stands for being censored.}

\item{W}{vector of treatment assignment. \code{W = 1} stands for treatment, and \code{W = 0} stands for control.}

\item{family}{a character string describing the error distribution and link function to be used in the model. \code{Family} can be one of the following options: "gaussian", "binomial", "poisson", "cox". Default is "gaussian".}

\item{prop}{method for estimating the propensity score. \code{Prop} can be a character string naming the estimator: "logistic regression", "gradient boosting". \code{Prop} can also be a function and \code{prop(X)} should output the estimated propensity scores for each row of the covariate matrix \code{X}. Default is "logistic regression".}

\item{paramsW}{list of parameters for propensity score estimation.}

\item{baseline}{method for estimating the natural parameter functions in the treatment and control groups. \code{Baseline} can be a character string naming the estimator: "regression", "gradient boosting". \code{Baseline} can also be a list with two members: \code{eta1} denotes the estimator of the natural parameter function in the treatment group,   \code{eta0} denotes the estimator of the natural parameter function in the control group. The functions should take in a covariate matrix \code{X} and output the estimated quantities for each row of the covariate matrix \code{X}. Default is "regression".}

\item{params}{list of parameters for the natural parameter function estimation.}

\item{k}{number of folds for cross-fitting. Default is 2.}

\item{index.resample}{vector of subject indices in the permuted sample. We permute the original dataset and each subject will appear exactly once in the permuted sample. The permutation is done to break inherent ordering in the original dataset (if there is any) and makes sure the resulted folds are exchangeable.}

\item{foldIndex}{list of fold memberships. The j-th member is a vector containing all subjects of the bootstrap resample assigned to the j-th fold. If \code{foldIndex = NULL}, no sample splitting is done and the estimation of nuisance functions and causal parts are both based on the full dataset.}

\item{link.function}{link function from the conditional mean to the natural parameter.}

\item{var.function}{variance function from the conditional mean to the variance function.}
}
\value{
an object of class \code{DINA}..
 \itemize{
 \item{estimator}{vector of estimated coefficients.}
 \item{Sigma}{the variance matrix of the coefficients.}
}
}
\description{
This function is a helper for the function \code{DINA}.
}
