% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DINA.R
\name{DINA}
\alias{DINA}
\title{DINA}
\usage{
DINA(
  data,
  family = "gaussian",
  prop = NULL,
  baseline = "gradient boosting",
  k = 2,
  crossFitting = TRUE,
  bootstrap = F,
  B = 1,
  full = F,
  weight.mechanism = "random"
)
}
\arguments{
\item{family}{"gaussian"}

\item{prop}{"logistic regression"}

\item{baseline}{"gradient boosting"}

\item{k}{2}

\item{crossFitting}{TRUE}

\item{bootstrap}{to bagging}

\item{B}{100}

\item{inference}{FALSE}

\item{...}{pass to inference}
}
\description{
This function implements DINA ---  an estimator of heterogeneous treatment effects for general responses. DINA makes use of existing machine learning tools and enjoys desirable statistical properties.
}
