% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DINA.R
\name{DINAEstimator.step1}
\alias{DINAEstimator.step1}
\title{DINAEstimator.step1 here!!!}
\usage{
DINAEstimator.step1(
  X,
  Y,
  W,
  family,
  prop,
  paramsW = NULL,
  baseline = "gradient boosting",
  params,
  link.function = NA,
  var.function = NA
)
}
\arguments{
\item{X}{matrix of covariates.}

\item{Y}{vector of responses.}

\item{W}{vector of treatment assignment.}

\item{family}{a character string describing the error distribution and link function to be used in the model. \code{Family} can be one of the following options: "gaussian", "binomial", "poisson", "cox". Default is "gaussian".}

\item{prop}{method for estimating the propensity score. \code{Prop} can be a character string naming the estimator: "logistic regression", "gradient boosting". \code{Prop} can also be a function and \code{prop(X)} should output the estimated propensity scores for each row of the covariate matrix \code{X}. Default is "logistic regression".}

\item{paramsW}{list of parameters for propensity score estimation.}

\item{baseline}{method for estimating the natural parameter functions in the treatment and control groups. \code{Baseline} can be a character string naming the estimator: "regression", "gradient boosting". \code{Baseline} can also be a list with two members: \code{eta1} denotes the estimator of the natural parameter function in the treatment group,   \code{eta0} denotes the estimator of the natural parameter function in the control group. The functions should take in a covariate matrix \code{X} and output the estimated quantities for each row of the covariate matrix \code{X}. Default is "regression".}

\item{params}{list of parameters for the natural parameter function estimation.}

\item{link.function}{link function from the conditional mean to the natural parameter.}

\item{var.function}{variance function from the conditional mean to the variance function.}
}
\value{
a list of data.
 \itemize{
 \item{a}{the estimated nuisance function a(x).}
 \item{nu}{the estimated nuisance function nu(x).}
}
}
\description{
This function implements the first step of DINA: estimation of the nuisance functions.
}
