% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{not.censoring.probability}
\alias{not.censoring.probability}
\title{not.censoring.probability}
\usage{
not.censoring.probability(
  X = NULL,
  time = NULL,
  status = NULL,
  eta = NULL,
  censorType = NULL,
  params = list()
)
}
\arguments{
\item{censorType}{type of censoring. Types of censoring can be one of the following options: NULL (no censoring), "exponential", "cox", "uniform", "fixed", "unknown". For censoring types except NULL and "unknown", we compute the probability of getting censored in closed form or numerically. If the type of censoring is "unknown" or "exponential", a classifier is trained with the response whether a unit is censored and all available covariates as predictors.}
}
\value{
a function takes in covariate values and outputs the probabilities of not getting censored.
}
\description{
This function estimates the probability of not getting censored.
}
