% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.DINA.R
\name{predict.DINA}
\alias{predict.DINA}
\title{predict.DINA}
\usage{
\method{predict}{DINA}(object, newdata, ...)
}
\arguments{
\item{object}{object of class "DINA".}

\item{newdata}{data frame or matrix of covariates with which to predict.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a vector of predictions.
}
\description{
This function computes the predicted heterogeneous treatment effect based on a DINA object.
}
