% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rDirichlet}
\alias{rDirichlet}
\title{Generate Dirichlet Random Variable}
\usage{
rDirichlet(deltas)
}
\arguments{
\item{deltas}{A \code{vector} of Dirichlet parameters.}
}
\value{
A \code{vector} from a Dirichlet.
}
\description{
Sample a Dirichlet random variable.
}
\author{
Steven Andrew Culpepper
}
