% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rmultinomial}
\alias{rmultinomial}
\title{Generate Multinomial Random Variable}
\usage{
rmultinomial(ps)
}
\arguments{
\item{ps}{A \code{vector} for the probability of each category.}
}
\value{
A \code{vector} from a multinomial with probability ps.
}
\description{
Sample a multinomial random variable for given probabilities.
}
\author{
Steven Andrew Culpepper
}
