% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{sample.split}
\alias{sample.split}
\title{sample.split}
\usage{
sample.split(weight, k)
}
\arguments{
\item{weight}{weight vector}

\item{k}{number of folds. Default is 2.}
}
\value{
a matrix of samples. Each column stands for a fold. Each sample is repeated for \code{weight} times.
}
\description{
This function splits weighted samples into folds with (approximately) equal weight sum.
}
