% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{split.uniform}
\alias{split.uniform}
\title{split.uniform}
\usage{
\method{split}{uniform}(n, k = 2, replace = F)
}
\arguments{
\item{n}{sample size.}

\item{k}{number of folds. Default is 2.}

\item{replace}{logical value indicating the whether the sampling should be with replacement. If \code{TRUE}, sampling will be with replacement. Default is \code{FALSE}.}
}
\value{
a list indicating the resulting folds.
 \itemize{
 \item{index.resample}{vector of subject indices in the permuted sample. We permute the original dataset and each subject will appear exactly once in the permuted sample. The permutation is done to break inherent ordering in the original dataset (if there is any) and makes sure the resulted folds are exchangeable.}
 \item{foldIndex}{list of fold memberships. The j-th member is a vector containing all subjects of the bootstrap resample assigned to the j-th fold.}
}
}
\description{
This function splits samples into folds with equal fold size.
}
\examples{
set.seed(318)
test = split.uniform(n, k)
test$index.resample[test$foldIndex[[1]]]

}
