% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_alpha}
\alias{update_alpha}
\title{Update attributes and latent class probabilities}
\usage{
update_alpha(Amat, Q, ss, gs, Y, PIs, ALPHAS, delta0)
}
\arguments{
\item{Amat}{A C by K \code{matrix} of latent classes.}

\item{Q}{A N by K \code{matrix} indicating which skills are required for which items.}

\item{ss}{A J \code{vector} of item slipping parameters.}

\item{gs}{A J \code{vector} of item guessing parameters.}

\item{Y}{A N by J \code{matrix} of observed responses.}

\item{PIs}{A C \code{vector} of latent class probabilities.}

\item{ALPHAS}{A N by K \code{matrix} of latent attributes.}

\item{delta0}{A J \code{vector} of Dirichlet prior parameters.}
}
\value{
A N by K \code{matrix} of attributes and a C \code{vector} of class probabilities.
}
\description{
Update attributes and latent class probabilities by sampling from full
conditional distribution.
}
\author{
Steven Andrew Culpepper
}
