% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_sg}
\alias{update_sg}
\title{Update item parameters}
\usage{
update_sg(Y, Q, ALPHAS, ss_old, as0, bs0, ag0, bg0)
}
\arguments{
\item{Y}{A N by J \code{matrix} of observed responses.}

\item{Q}{A N by K \code{matrix} indicating which skills are required for which items.}

\item{ALPHAS}{A N by K \code{matrix} of latent attributes.}

\item{ss_old}{A J \code{vector} of item slipping parameters from prior iteration.}

\item{as0}{Slipping prior alpha parameter for Beta distribution.}

\item{bs0}{Slipping prior beta parameter for Beta distribution.}

\item{ag0}{Guessing prior alpha parameter for Beta distribution.}

\item{bg0}{Guessing prior beta parameter for Beta distribution.}
}
\value{
A list with two J \code{vectors} of guessing and slipping parameters.
}
\description{
Update guessing and slipping parameters from full conditional distribution.
}
\author{
Steven Andrew Culpepper
}
