\name{dina-package}
\alias{dina-package}
\alias{dina}
\docType{package}
\title{
Bayesian Estimation of DINA Model
}
\description{
Estimate the Deterministic Input, Noisy ``And'' Gate (DINA) cognitive diagnostic model parameters using the Gibbs sampler described by Culpepper (2015).
}
\details{
\tabular{ll}{
Package: \tab dina\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Date: \tab 2015-02-25\cr
License: \tab GPL (>= 2)\cr
}

}
\author{
Steven Andrew Culpepper
Maintainer: Steven Andrew Culpepper <sculpepp@illinois.edu>
}
\references{
Culpepper (2015). Bayesian estimation of the DINA model with Gibbs sampling. Journal of Educational and Behavioral Statistics. DOI: 10.3102/1076998615595403 
}