% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dina-est.R
\name{DINA_Gibbs}
\alias{DINA_Gibbs}
\title{Deprecated functions}
\usage{
DINA_Gibbs(...)
}
\arguments{
\item{...}{Old parameters}
}
\description{
Functions found within this help documentation have been deprecated.
}
\details{
Deprecated functions
\itemize{
\item \code{DINA_Gibbs} in favor of \code{dina}
}
}
