% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.DINA.R
\name{predict.DINA}
\alias{predict.DINA}
\title{DINA}
\usage{
\method{predict}{DINA}(object, ..., interval = "none", sd.method = "bootstrap")
}
\arguments{
\item{sd.method}{"bootstrap", "approximation".}

\item{newdata}{the data frame in which to look for variables with which to predict. If omitted, the fitted values are used. Default is NULL.}

\item{interval:}{"none", "confidence", "prediction?". Type of interval calculation. If omitted, point estimates are provided. Default is NULL.}

\item{level}{confidence level. Default is 0.95.}
}
\description{
This function adapts the predict.lm for DINA.
}
