% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dina-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{sim_dina}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{simcdm}{\code{\link[simcdm]{sim_dina}}}
}}

